requirements.txt
astroid==2.15.6
asttokens==2.2.1
autopep8==2.0.4
backcall==0.2.0
cffi==1.15.1
colorama==0.4.6
comm==0.1.4
cryptography==41.0.3
debugpy==1.6.7.post1
decorator==5.1.1
dill==0.3.7
et-xmlfile==1.1.0
executing==1.2.0
importlib-metadata==6.8.0
ipykernel==6.25.1
ipython==8.14.0
isort==5.12.0
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.3.1
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
nest-asyncio==1.5.7
numpy==1.25.2
openpyxl==3.1.2
oracledb==1.4.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pyarrow==13.0.0
pycodestyle==2.11.0
pycparser==2.21
Pygments==2.16.1
pylance==0.6.5
pylint==2.17.5
pyodbc==4.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
pyzmq==25.1.1
schedule==1.2.0
six==1.16.0
stack-data==0.6.2
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
wcwidth==0.2.6
wrapt==1.15.0
zipp==3.16.2
